Confirmed intial transfer worked. Needed to have separate package.json files for each folder (node server vs react server). Will do some additional edits on this branch until confirming all features work as expected
