+++ .gitignore
@@ -1,7 +1,8 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
+/react/node_modules
+/node/node_modules
 /.pnp
 .pnp.js
 
@@ -12,6 +13,7 @@
 /build
 
 # misc
+/react/path.env
 .DS_Store
 .env.local
 .env.development.local
@@ -21,3 +23,6 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+/notes
+.gitignore
