+++ react/package.json
@@ -25,7 +25,7 @@
     "web-vitals": "^2.1.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-scripts start $PUBLIC_URL",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
