+++ react/src/components/chart chartist/form-data/BarChartFormData.js
