+++ react/src/components/chart chartist/form-data/LineChartFormData.js
