+++ react/src/components/chart chartist/forms/ChartSection.js
