+++ react/src/components/chart chartist/forms/ChartSettings.js
