+++ react/src/components/chart d3/BarChart.js
