+++ react/src/components/chart d3/forms/BarChartForm.js
